refactor(auth): restructure authentication components and utilities
- Extract ProtectedComponent into separate file
- Create dedicated utils.tsx for auth utility functions
- Improve login flow with better environment handling
- Refactor redirectToAuth logic for better maintainability
- Move LOGIN_TYPE constant to env.ts for centralized configuration
- Update imports and exports to reflect new structure
