+++ contexts/Auth/index.ts
@@ -1,2 +1,4 @@
-export * from './AuthContext';
 export { RoleEnum } from '@/services/users';
+export { default as ProtectedComponent } from './ProtectedComponent';
+export * from './AuthContext';
+export * from './utils';
